requirements.txt
torch
numpy
transformers
sentence-transformers
faiss-cpu
duckduckgo-search
keybert
rapidfuzz
bitsandbytes
regex
